refactor: move composer playback scenarios
