+++ src/loushang/coding/ui/playback_runner.py
@@ -1,14 +1,12 @@
 from __future__ import annotations
 
 import argparse
-import asyncio
 import json
 import sys
 from collections.abc import Sequence
 from pathlib import Path
 from typing import TextIO
 
-from loushang.coding.ui.completion import coding_inline_completion_provider
 from loushang.coding.ui.perf_probe import build_synthetic_long_transcript_records
 from loushang.coding.ui.playback import (
     NativeTuiInputPlaybackResult,
@@ -25,9 +23,6 @@ from loushang.coding.ui.playback_fakes import (
     RecordingTerminalMode as _RecordingTerminalMode,
 )
 from loushang.coding.ui.playback_fakes import (
-    SessionCommandPlaybackSession as _SessionCommandSession,
-)
-from loushang.coding.ui.playback_fakes import (
     recording_drain as _recording_drain,
 )
 from loushang.coding.ui.playback_scenarios.budgets import (
@@ -35,6 +30,7 @@ from loushang.coding.ui.playback_scenarios.budgets import (
     LONG_TRANSCRIPT_FRAME_BUDGET,
 )
 from loushang.coding.ui.playback_scenarios.command import COMMAND_ROUTING_SCENARIOS
+from loushang.coding.ui.playback_scenarios.composer import COMPOSER_SCENARIOS
 from loushang.coding.ui.playback_scenarios.lifecycle import LIFECYCLE_SCENARIOS
 from loushang.coding.ui.playback_scenarios.surface import SURFACE_SCENARIOS
 from loushang.coding.ui.playback_suite import (
@@ -164,86 +160,6 @@ def _json_summary(results: Sequence[NativePlaybackScenarioResult]) -> dict[str, 
     }
 
 
-def _run_completion_tab() -> NativeTuiInputPlaybackResult:
-    result = (
-        NativeTuiInputScenario(width=80, height=12)
-        .with_completion_items("/model", "/models")
-        .render()
-        .type_text("/mod")
-        .tab()
-        .run()
-    )
-    result.assert_composer_text("/model ")
-    result.assert_visible_contains("› /model")
-    result.assert_no_clear_screen()
-    result.assert_cursor_matches_diagnostics()
-    INTERACTION_FRAME_BUDGET.assert_result(result, skip_first=True)
-    return result
-
-
-def _run_completion_session_command() -> NativeTuiInputPlaybackResult:
-    session = _SessionCommandSession()
-    scenario = NativeTuiInputScenario(width=80, height=12)
-    scenario.app.composer.set_completion_provider(asyncio.run(coding_inline_completion_provider(session)))
-
-    result = scenario.render().type_text("/na").tab().run()
-
-    result.assert_composer_text("/name ")
-    result.assert_visible_contains("› /name")
-    result.assert_no_clear_screen()
-    result.assert_cursor_matches_diagnostics()
-    assert session.commands == []
-    assert session.prompts == []
-    INTERACTION_FRAME_BUDGET.assert_result(result, skip_first=True)
-    return result
-
-
-def _run_completion_navigation_priority() -> NativeTuiInputPlaybackResult:
-    result = (
-        NativeTuiInputScenario(width=80, height=12)
-        .with_history("history prompt")
-        .with_completion_items("/model", "/models")
-        .render()
-        .type_text("/mod")
-        .key("\x1b[B")
-        .tab()
-        .run()
-    )
-    result.assert_composer_text("/models ")
-    result.assert_visible_contains("› /models")
-    result.assert_no_clear_screen()
-    result.assert_cursor_matches_diagnostics()
-    INTERACTION_FRAME_BUDGET.assert_result(result, skip_first=True)
-    return result
-
-
-def _run_history_navigation() -> NativeTuiInputPlaybackResult:
-    result = (
-        NativeTuiInputScenario(width=80, height=12)
-        .with_history("first prompt", "second prompt")
-        .render()
-        .type_text("draft")
-        .key("\x1b[A")
-        .key("\x1b[A")
-        .key("\x1b[B")
-        .key("\x1b[B")
-        .run()
-    )
-    assert [state["composer_text"] for state in result.step_coding_states[1:]] == [
-        "draft",
-        "second prompt",
-        "first prompt",
-        "second prompt",
-        "draft",
-    ]
-    result.assert_composer_text("draft")
-    result.assert_visible_contains("› draft")
-    result.assert_no_clear_screen()
-    result.assert_cursor_matches_diagnostics()
-    INTERACTION_FRAME_BUDGET.assert_result(result, skip_first=True)
-    return result
-
-
 def _run_local_command() -> NativeTuiInputPlaybackResult:
     result = (
         NativeTuiInputScenario(width=80, height=12)
@@ -327,22 +243,6 @@ def _run_tool_output_preview() -> NativeTuiInputPlaybackResult:
     return result
 
 
-def _run_bracketed_paste_large_marker() -> NativeTuiInputPlaybackResult:
-    pasted = "\n".join(f"line {index}" for index in range(10))
-    result = (
-        NativeTuiInputScenario(width=80, height=12)
-        .render()
-        .key(f"{BRACKETED_PASTE_START}{pasted}{BRACKETED_PASTE_END}")
-        .run()
-    )
-    result.assert_composer_text(pasted)
-    result.assert_visible_contains("[paste #1 +10 lines]")
-    result.assert_no_clear_screen()
-    result.assert_cursor_matches_diagnostics()
-    INTERACTION_FRAME_BUDGET.assert_result(result, skip_first=True)
-    return result
-
-
 def _run_native_loop_split_bracketed_paste() -> object:
     playback = NativeTuiLoopPlayback(width=80, height=12)
     pasted = "alpha\nbeta\ngamma"
@@ -368,60 +268,6 @@ def _run_native_loop_split_bracketed_paste() -> object:
     result.assert_text_contains("› alpha")
     result.assert_text_contains("split paste submitted once")
     result.assert_no_clear_screen()
-    return result
-
-
-def _run_resize_reflow_stable() -> NativeTuiInputPlaybackResult:
-    result = (
-        NativeTuiInputScenario(width=80, height=12)
-        .render()
-        .type_text("resize keeps composer stable")
-        .resize(width=42, height=8)
-        .type_text(" after shrink")
-        .resize(width=100, height=14)
-        .type_text(" after grow")
-        .run()
-    )
-    result.assert_composer_text("resize keeps composer stable after shrink after grow")
-    result.assert_visible_contains("after grow")
-    assert any(step.diagnostics.operation_class == "resize_repaint" for step in result.steps)
-    result.assert_no_clear_scrollback()
-    result.assert_cursor_matches_diagnostics()
-    INTERACTION_FRAME_BUDGET.assert_result(result, skip_first=True)
-    return result
-
-
-def _run_wide_char_input_cursor() -> NativeTuiInputPlaybackResult:
-    result = (
-        NativeTuiInputScenario(width=32, height=10)
-        .render()
-        .type_chars("你好🙂 terminal")
-        .run()
-    )
-    result.assert_composer_text("你好🙂 terminal")
-    result.assert_visible_contains("你好🙂 terminal")
-    result.assert_no_clear_screen()
-    result.assert_cursor_matches_diagnostics()
-    INTERACTION_FRAME_BUDGET.assert_result(result, skip_first=True)
-    return result
-
-
-def _run_keyboard_shift_enter_newline() -> NativeTuiInputPlaybackResult:
-    result = (
-        NativeTuiInputScenario(width=80, height=12)
-        .render()
-        .type_text("first line")
-        .key("\x1b[13;2u")
-        .type_text("second line")
-        .enter()
-        .run()
-    )
-    result.assert_prompt_texts("first line\nsecond line")
-    result.assert_composer_text("")
-    result.assert_visible_contains("› first line")
-    result.assert_no_clear_screen()
-    result.assert_cursor_matches_diagnostics()
-    INTERACTION_FRAME_BUDGET.assert_result(result, skip_first=True)
     return result
 
 
@@ -583,27 +429,7 @@ def _run_apple_shift_enter_normalized() -> object:
 
 DEFAULT_SUITE = NativePlaybackSuite(
     (
-        NativePlaybackScenarioSpec(
-            name="completion-tab",
-            description="Apply tab completion without clearing or repainting the screen.",
-            run=_run_completion_tab,
-        ),
-        NativePlaybackScenarioSpec(
-            name="completion-session-command",
-            description="Apply session command completion without executing the selected command.",
-            run=_run_completion_session_command,
-            tags=("completion", "command", "session"),
-        ),
-        NativePlaybackScenarioSpec(
-            name="completion-navigation-priority",
-            description="Route completion navigation before history navigation.",
-            run=_run_completion_navigation_priority,
-        ),
-        NativePlaybackScenarioSpec(
-            name="history-navigation",
-            description="Browse prompt history from a non-empty draft and restore the draft.",
-            run=_run_history_navigation,
-        ),
+        *COMPOSER_SCENARIOS,
         NativePlaybackScenarioSpec(
             name="local-command",
             description="Route a local command without echoing it as a prompt.",
@@ -629,29 +455,9 @@ DEFAULT_SUITE = NativePlaybackSuite(
             run=_run_tool_output_preview,
         ),
         NativePlaybackScenarioSpec(
-            name="bracketed-paste-large-marker",
-            description="Render a large bracketed paste as a stable composer marker.",
-            run=_run_bracketed_paste_large_marker,
-        ),
-        NativePlaybackScenarioSpec(
             name="native-loop-split-bracketed-paste",
             description="Keep split native bracketed paste atomic until the end marker arrives.",
             run=_run_native_loop_split_bracketed_paste,
-        ),
-        NativePlaybackScenarioSpec(
-            name="resize-reflow-stable",
-            description="Keep composer text and cursor stable across terminal resizes.",
-            run=_run_resize_reflow_stable,
-        ),
-        NativePlaybackScenarioSpec(
-            name="wide-char-input-cursor",
-            description="Keep CJK and emoji input cursor diagnostics aligned.",
-            run=_run_wide_char_input_cursor,
-        ),
-        NativePlaybackScenarioSpec(
-            name="keyboard-shift-enter-newline",
-            description="Route raw Shift+Enter to composer newline before submission.",
-            run=_run_keyboard_shift_enter_newline,
         ),
         NativePlaybackScenarioSpec(
             name="mouse-select-active-surface",
